centralized handling input

+++ actions/delete.php
@@ -4,10 +4,10 @@ require "../config/database.php";
 require "../includes/functions.php";
 
 $id = $_REQUEST['id'];
-
+$user_id = $_SESSION['user_id'];
 try {
-    $stmt = $dbh->prepare("DELETE FROM books WHERE id=?");
-    $stmt->execute([$id]);
+    $stmt = $dbh->prepare("DELETE FROM books WHERE id=? AND user_id=?");
+    $stmt->execute([$id, $user_id]);
     redirect();
 } catch (PDOException $e) {
     die("Database Error: " . $e->getMessage());
